+++ CHANGELOG.md
@@ -10,6 +10,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Added TLV parser for `IBILteCellInfoT`
 
+### Changes
+
+- tools: Made asString extractor more fail-safe when resolving pointers
+
 ## [v2.0.0](https://github.com/seemoo-lab/aristoteles/tree/v2.0.0) (28-11-2021)
 
 ### Features